DST-1263 refactor the acs controller into smaller files
